Introduce liveCoreClrBuildConfig and liveLibrariesBuildConfig
These complementary properties should let us drop all automatic
transformations of CoreCLR vs. library build modes and thus
substantially simplify the live-live pipeline and make it more
general at the same time. I have reverted the changes to the
build-test scripts that were taking care of this logic on the
local build side as well as the change to
eng/pipelines/libraries/base-job and bits of the other YAML files
on the CI side.

Thanks

Tomas